tweaked some things for online deployment

+++ public/src/components/Logout.jsx
@@ -3,12 +3,19 @@ import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import { BiPowerOff } from "react-icons/bi";
+import { logoutRoute } from "../utils/APIRoutes";
 
 export default function Logout() {
   const navigate = useNavigate();
   const handleClick = async () => {
-    localStorage.clear();
-    navigate("/login");
+    const id = await JSON.parse(
+      localStorage.getItem(process.env.REACT_APP_LOCALHOST_KEY)
+    )._id;
+    const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
+    }
   };
   return (
     <Button onClick={handleClick}>
